on post mdoel added pending tests for validations and associations on
One easy thing to test is validations.  This app currently does not have validations, which may be a problem.

Assocations are slightly harder to test, and according to Kori nobody ever specifically tests if their associations are working because that happens by default when you do integration tests.

I think it would be valuable to prove to ourselves how the polymorphic association works, but that may be a rabbit hole.

Mission for the next set of tests: get basic 'does this attribute exist in the model' tests.